Commit work on prototype designed to figure out flow
The goal of this prototype was to figure out whether we should let the
user edit their own database, or take them to a separate database.

This isn't really so much a "prototype", as it is a mockup of the UI
together with a snippet of what the text and behavior may look like.

I decided on a flow where the user edits their own database, but we show
examples (the images) of what we expect Thinktool to be used for.